% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NestedMenu.R
\name{NestedMenu}
\alias{NestedMenu}
\title{'Nested menu' widget}
\usage{
NestedMenu(
  label,
  items,
  trigger = "left",
  style = "primary",
  size = NULL,
  elementId = NULL
)
}
\arguments{
\item{label}{the label of the root button}

\item{items}{list of items for the nested menu; see the
\link[=NestedMenu-shiny]{Shiny example}}

\item{trigger}{the way the menu is triggered: \code{"left"} to trigger
on a left-click, \code{"right"} to trigger on a right-click,
\code{"hover"} to trigger on hover}

\item{style}{a Bootstrap style for the root button: \code{"primary"}, 
\code{"info"}, \code{"success"}, \code{"warning"} or \code{"danger"}}

\item{size}{size the root button: \code{NULL} (normal), \code{"lg"} (large),
\code{"sm"} (small) or \code{"xs"} (extra-small) 
\code{"info"}, \code{"success"}, \code{"warning"} or \code{"danger"}}

\item{elementId}{a HTML id; this is usually useless}
}
\description{
'Nested menu' HTML widget.
}
